refactor: 도메인 뷰 컨트롤러 패키지 구조 추가
