+++ src/test/java/lotto/LottoNumberGeneratorTest.java
@@ -1,5 +1,7 @@
 package lotto;
 
+import lotto.domain.FixedLottoNumberGenerator;
+import lotto.domain.RandomLottoNumberGenerator;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
